Defining initial endpoints and requirements
This commit creates the initial README which includes the overall description
of the project and an initial pass of the rest endpoints.
It also includes the requirements necessary for running or accessing the server,
these are expected to be extracted out later and turned into an auto setup script
for the server.